+++ visqol-rs/src/constants.rs
@@ -1,4 +1,4 @@
-pub const DURATION_MISMATCH_TOLERANCE: f64 = 1.0;
+pub const DURATION_MISMATCH_TOLERANCE: f32 = 1.0;
 
 pub const NUM_BANDS_SPEECH: usize = 21;
 pub const NUM_BANDS_AUDIO: usize = 32;
@@ -6,8 +6,8 @@ pub const DEFAULT_WINDOW_SIZE: usize = 32;
 
 pub const PATCH_SIZE_AUDIO: usize = 30;
 pub const PATCH_SIZE_SPEECH: usize = 20;
-pub const MINIMUM_FREQ: f64 = 50.0;
-pub const OVERLAP: f64 = 0.25;
-pub const WINDOW_DURATION: f64 = 0.08;
+pub const MINIMUM_FREQ: f32 = 50.0;
+pub const OVERLAP: f32 = 0.25;
+pub const WINDOW_DURATION: f32 = 0.08;
 
 pub const NUM_FILTER_CONDITIONS: usize = 2;
