+++ visqol-rs/src/patch_similarity_comparator.rs
@@ -5,30 +5,30 @@ use serde::Serialize;
 /// The term `Patch` here refers to a single of spectrogram data produced by a PatchCreator)
 pub struct PatchSimilarityResult {
     /// Means of the individual frequency bands
-    pub freq_band_means: Vec<f64>,
+    pub freq_band_means: Vec<f32>,
     /// Standard deviation of the individual frequency bands
-    pub freq_band_stddevs: Vec<f64>,
+    pub freq_band_stddevs: Vec<f32>,
     /// Energy of the degraded file per frequency band
-    pub freq_band_deg_energy: Vec<f64>,
+    pub freq_band_deg_energy: Vec<f32>,
     /// Calculated patch similarity score
-    pub similarity: f64,
+    pub similarity: f32,
     /// Reference start of patch in seconds
-    pub ref_patch_start_time: f64,
+    pub ref_patch_start_time: f32,
     /// Reference end of patch in seconds
-    pub ref_patch_end_time: f64,
+    pub ref_patch_end_time: f32,
     /// Degraded start of patch in seconds
-    pub deg_patch_start_time: f64,
+    pub deg_patch_start_time: f32,
     /// Degraded end of patch in seconds
-    pub deg_patch_end_time: f64,
+    pub deg_patch_end_time: f32,
 }
 
 impl PatchSimilarityResult {
     /// Creates a new similarity result, stores mean, std and energy of degraded signal and sets the time information to 0
     pub fn new(
-        freq_band_means: Vec<f64>,
-        freq_band_stddevs: Vec<f64>,
-        freq_band_deg_energy: Vec<f64>,
-        similarity: f64,
+        freq_band_means: Vec<f32>,
+        freq_band_stddevs: Vec<f32>,
+        freq_band_deg_energy: Vec<f32>,
+        similarity: f32,
     ) -> Self {
         Self {
             freq_band_means,
@@ -46,9 +46,9 @@ impl PatchSimilarityResult {
 impl Default for PatchSimilarityResult {
     fn default() -> Self {
         Self {
-            freq_band_means: Vec::<f64>::new(),
-            freq_band_stddevs: Vec::<f64>::new(),
-            freq_band_deg_energy: Vec::<f64>::new(),
+            freq_band_means: Vec::<f32>::new(),
+            freq_band_stddevs: Vec::<f32>::new(),
+            freq_band_deg_energy: Vec::<f32>::new(),
             similarity: 0.0,
             ref_patch_start_time: 0.0,
             ref_patch_end_time: 0.0,
@@ -62,7 +62,7 @@ impl Default for PatchSimilarityResult {
 pub trait PatchSimilarityComparator {
     fn measure_patch_similarity(
         &self,
-        ref_patch: &mut ImagePatch<f64>,
-        deg_patch: &mut ImagePatch<f64>,
+        ref_patch: &mut ImagePatch<f32>,
+        deg_patch: &mut ImagePatch<f32>,
     ) -> PatchSimilarityResult;
 }
